+++ examples/basic/basic-example.ts
@@ -97,7 +97,8 @@ async function main() {
     index,
     utils.convertDecimalToNativeInteger(0.1),
     utils.convertDecimalToNativeLotSize(2),
-    types.Side.BID
+    types.Side.BID,
+    { tifOptions: {}, orderType: types.OrderType.LIMIT } // Extra optional parameters
   );
 
   // See our order in the orderbook.