+++ src/program-instructions.ts
@@ -186,6 +186,8 @@ export function withdrawIx(
       tokenProgram: TOKEN_PROGRAM_ID,
       greeks: subExchange.zetaGroup.greeks,
       oracle: subExchange.zetaGroup.oracle,
+      oracleBackupFeed: subExchange.zetaGroup.oracleBackupFeed,
+      oracleBackupProgram: constants.CHAINLINK_PID,
       socializedLossAccount: subExchange.socializedLossAccountAddress,
     },
   });
@@ -195,6 +197,7 @@ export async function initializeOpenOrdersIx(
   asset: Asset,
   market: PublicKey,
   userKey: PublicKey,
+  authority: PublicKey,
   marginAccount: PublicKey
 ): Promise<[TransactionInstruction, PublicKey]> {
   const [openOrdersPda, _openOrdersNonce] = await utils.getOpenOrders(
@@ -217,8 +220,8 @@ export async function initializeOpenOrdersIx(
         systemProgram: SystemProgram.programId,
         openOrders: openOrdersPda,
         marginAccount: marginAccount,
-        authority: userKey,
-        payer: userKey,
+        authority: authority,
+        payer: authority,
         market: market,
         rent: SYSVAR_RENT_PUBKEY,
         serumAuthority: Exchange.serumAuthority,
@@ -323,6 +326,8 @@ export function placeOrderV3Ix(
           pcWallet: marketData.quoteVault,
         },
         oracle: subExchange.zetaGroup.oracle,
+        oracleBackupFeed: subExchange.zetaGroup.oracleBackupFeed,
+        oracleBackupProgram: constants.CHAINLINK_PID,
         marketNode: subExchange.greeks.nodeKeys[marketIndex],
         marketMint:
           side == types.Side.BID
@@ -404,6 +409,8 @@ export function placeOrderV4Ix(
           pcWallet: marketData.quoteVault,
         },
         oracle: subExchange.zetaGroup.oracle,
+        oracleBackupFeed: subExchange.zetaGroup.oracleBackupFeed,
+        oracleBackupProgram: constants.CHAINLINK_PID,
         marketNode: subExchange.greeks.nodeKeys[marketIndex],
         marketMint:
           side == types.Side.BID
@@ -483,6 +490,8 @@ export function placePerpOrderIx(
           pcWallet: marketData.quoteVault,
         },
         oracle: subExchange.zetaGroup.oracle,
+        oracleBackupFeed: subExchange.zetaGroup.oracleBackupFeed,
+        oracleBackupProgram: constants.CHAINLINK_PID,
         marketMint:
           side == types.Side.BID
             ? marketData.serumMarket.quoteMintAddress
@@ -564,6 +573,8 @@ export function placePerpOrderV2Ix(
           pcWallet: marketData.quoteVault,
         },
         oracle: subExchange.zetaGroup.oracle,
+        oracleBackupFeed: subExchange.zetaGroup.oracleBackupFeed,
+        oracleBackupProgram: constants.CHAINLINK_PID,
         marketMint:
           side == types.Side.BID
             ? marketData.serumMarket.quoteMintAddress
@@ -803,6 +814,8 @@ export function forceCancelOrderByOrderIdIx(
       accounts: {
         greeks: subExchange.zetaGroup.greeks,
         oracle: subExchange.zetaGroup.oracle,
+        oracleBackupFeed: subExchange.zetaGroup.oracleBackupFeed,
+        oracleBackupProgram: constants.CHAINLINK_PID,
         cancelAccounts: {
           zetaGroup: subExchange.zetaGroupAddress,
           state: Exchange.stateAddress,
@@ -832,6 +845,8 @@ export function forceCancelOrdersIx(
     accounts: {
       greeks: subExchange.zetaGroup.greeks,
       oracle: subExchange.zetaGroup.oracle,
+      oracleBackupFeed: subExchange.zetaGroup.oracleBackupFeed,
+      oracleBackupProgram: constants.CHAINLINK_PID,
       cancelAccounts: {
         zetaGroup: subExchange.zetaGroupAddress,
         state: Exchange.stateAddress,
@@ -1032,6 +1047,8 @@ export async function initializeZetaGroupIx(
   asset: Asset,
   underlyingMint: PublicKey,
   oracle: PublicKey,
+  oracleBackupFeed: PublicKey,
+  oracleBackupProgram: PublicKey,
   pricingArgs: InitializeZetaGroupPricingArgs,
   perpArgs: UpdatePerpParametersArgs,
   marginArgs: UpdateMarginParametersArgs,
@@ -1127,6 +1144,8 @@ export async function initializeZetaGroupIx(
         underlyingMint,
         zetaProgram: Exchange.programId,
         oracle,
+        oracleBackupFeed,
+        oracleBackupProgram,
         zetaGroup,
         greeks,
         perpSyncQueue,
@@ -1215,6 +1234,8 @@ export function liquidateIx(
       liquidatorMarginAccount,
       greeks: subExchange.zetaGroup.greeks,
       oracle: subExchange.zetaGroup.oracle,
+      oracleBackupFeed: subExchange.zetaGroup.oracleBackupFeed,
+      oracleBackupProgram: constants.CHAINLINK_PID,
       market,
       zetaGroup: subExchange.zetaGroupAddress,
       liquidatedMarginAccount,
@@ -1288,6 +1309,8 @@ export function retreatMarketNodesIx(
       zetaGroup: subExchange.zetaGroupAddress,
       greeks: subExchange.greeksAddress,
       oracle: subExchange.zetaGroup.oracle,
+      oracleBackupFeed: subExchange.zetaGroup.oracleBackupFeed,
+      oracleBackupProgram: constants.CHAINLINK_PID,
     },
     remainingAccounts,
   });
@@ -1305,6 +1328,8 @@ export function updatePricingIx(
       zetaGroup: subExchange.zetaGroupAddress,
       greeks: subExchange.greeksAddress,
       oracle: subExchange.zetaGroup.oracle,
+      oracleBackupFeed: subExchange.zetaGroup.oracleBackupFeed,
+      oracleBackupProgram: constants.CHAINLINK_PID,
       perpMarket: marketData.address,
       perpBids: subExchange.markets.perpMarket.serumMarket.bidsAddress,
       perpAsks: subExchange.markets.perpMarket.serumMarket.asksAddress,
@@ -1508,6 +1533,8 @@ export function initializeMarketStrikesIx(
       state: Exchange.stateAddress,
       zetaGroup: subExchange.zetaGroupAddress,
       oracle: subExchange.zetaGroup.oracle,
+      oracleBackupFeed: subExchange.zetaGroup.oracleBackupFeed,
+      oracleBackupProgram: constants.CHAINLINK_PID,
     },
   });
 }
@@ -2066,6 +2093,8 @@ export function positionMovementIx(
   user: PublicKey,
   greeks: PublicKey,
   oracle: PublicKey,
+  oracleBackupFeed: PublicKey,
+  oracleBackupProgram: PublicKey,
   movementType: types.MovementType,
   movements: PositionMovementArg[]
 ): TransactionInstruction {
@@ -2081,6 +2110,8 @@ export function positionMovementIx(
         authority: user,
         greeks,
         oracle,
+        oracleBackupFeed,
+        oracleBackupProgram,
       },
     }
   );
@@ -2234,6 +2265,23 @@ export async function toggleMarketMakerIx(
       state: Exchange.stateAddress,
       admin: Exchange.state.admin,
       marginAccount,
+    },
+  });
+}
+
+export function editDelegatedPubkeyIx(
+  asset: Asset,
+  delegatedPubkey: PublicKey,
+  marginAccount: PublicKey,
+  authority: PublicKey
+): TransactionInstruction {
+  return Exchange.program.instruction.editDelegatedPubkey(delegatedPubkey, {
+    accounts: {
+      state: Exchange.stateAddress,
+      zetaGroup: Exchange.getZetaGroupAddress(asset),
+      marginAccount,
+      tokenProgram: TOKEN_PROGRAM_ID,
+      authority,
     },
   });
 }